# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/reconciliations:
    get:
      tags:
        - Reconciliations
      summary: List Reconciliations
      description: Get a paginated list of Reconciliations with optional filtering and sorting
      operationId: listReconciliations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Reconciliations
      summary: Create Reconciliation
      description: Create a new Reconciliation
      operationId: createReconciliation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReconciliationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reconciliation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/partial_reconciles:
    get:
      tags:
        - PartialReconciles
      summary: List PartialReconciles
      description: Get a paginated list of PartialReconciles with optional filtering and sorting
      operationId: listPartialReconciles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartialReconcileList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PartialReconciles
      summary: Create PartialReconcile
      description: Create a new PartialReconcile
      operationId: createPartialReconcile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartialReconcileRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartialReconcile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Reconciliation:
      type: object
      required:
        - id
        - company_id
        - reconciliation_number
        - account_id
        - account_number
        - account_name
        - reconciliation_type
        - period_start
        - period_end
        - statement_date
        - opening_book_balance
        - opening_statement_balance
        - closing_book_balance
        - closing_statement_balance
        - total_matched_debits
        - total_matched_credits
        - matched_count
        - total_unmatched_book_debits
        - total_unmatched_book_credits
        - unmatched_book_count
        - total_unmatched_statement_debits
        - total_unmatched_statement_credits
        - unmatched_statement_count
        - outstanding_deposits
        - outstanding_checks
        - deposits_in_transit
        - bank_charges
        - bank_interest
        - nsf_checks
        - other_adjustments
        - difference
        - is_balanced
        - status
        - has_adjusting_entries
        - adjusting_journal_ids
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        reconciliation_number:
          type: string
          maxLength: 30
        account_id:
          type: string
          format: uuid
        account_number:
          type: string
          maxLength: 20
        account_name:
          type: string
          maxLength: 100
        reconciliation_type:
          type: string
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        statement_date:
          type: string
          format: date
        opening_book_balance:
          type: string
        opening_statement_balance:
          type: string
        previous_reconciliation_id:
          nullable: true
          type: string
          format: uuid
        closing_book_balance:
          type: string
        closing_statement_balance:
          type: string
        total_matched_debits:
          type: string
        total_matched_credits:
          type: string
        matched_count:
          type: integer
          format: int32
        total_unmatched_book_debits:
          type: string
        total_unmatched_book_credits:
          type: string
        unmatched_book_count:
          type: integer
          format: int32
        total_unmatched_statement_debits:
          type: string
        total_unmatched_statement_credits:
          type: string
        unmatched_statement_count:
          type: integer
          format: int32
        outstanding_deposits:
          type: string
        outstanding_checks:
          type: string
        deposits_in_transit:
          type: string
        bank_charges:
          type: string
        bank_interest:
          type: string
        nsf_checks:
          type: string
        other_adjustments:
          type: string
        adjusted_book_balance:
          nullable: true
          type: string
        adjusted_statement_balance:
          nullable: true
          type: string
        difference:
          type: string
        is_balanced:
          type: boolean
        status:
          type: string
        started_at:
          nullable: true
          type: string
          format: date-time
        started_by:
          nullable: true
          type: string
          format: uuid
        completed_at:
          nullable: true
          type: string
          format: date-time
        completed_by:
          nullable: true
          type: string
          format: uuid
        reviewed_at:
          nullable: true
          type: string
          format: date-time
        reviewed_by:
          nullable: true
          type: string
          format: uuid
        has_adjusting_entries:
          type: boolean
        adjusting_journal_ids:
          type: object
        statement_source:
          nullable: true
          type: string
          maxLength: 50
        statement_file_url:
          nullable: true
          type: string
          maxLength: 500
        import_date:
          nullable: true
          type: string
          format: date-time
        notes:
          nullable: true
          type: string
          maxLength: 1000
        discrepancy_notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreateReconciliationRequest:
      type: object
      required:
        - company_id
        - reconciliation_number
        - account_id
        - account_number
        - account_name
        - reconciliation_type
        - period_start
        - period_end
        - statement_date
        - opening_book_balance
        - opening_statement_balance
        - closing_book_balance
        - closing_statement_balance
        - total_matched_debits
        - total_matched_credits
        - matched_count
        - total_unmatched_book_debits
        - total_unmatched_book_credits
        - unmatched_book_count
        - total_unmatched_statement_debits
        - total_unmatched_statement_credits
        - unmatched_statement_count
        - outstanding_deposits
        - outstanding_checks
        - deposits_in_transit
        - bank_charges
        - bank_interest
        - nsf_checks
        - other_adjustments
        - difference
        - is_balanced
        - status
        - has_adjusting_entries
        - adjusting_journal_ids
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        reconciliation_number:
          type: string
          maxLength: 30
        account_id:
          type: string
          format: uuid
        account_number:
          type: string
          maxLength: 20
        account_name:
          type: string
          maxLength: 100
        reconciliation_type:
          type: string
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        statement_date:
          type: string
          format: date
        opening_book_balance:
          type: string
        opening_statement_balance:
          type: string
        previous_reconciliation_id:
          nullable: true
          type: string
          format: uuid
        closing_book_balance:
          type: string
        closing_statement_balance:
          type: string
        total_matched_debits:
          type: string
        total_matched_credits:
          type: string
        matched_count:
          type: integer
          format: int32
        total_unmatched_book_debits:
          type: string
        total_unmatched_book_credits:
          type: string
        unmatched_book_count:
          type: integer
          format: int32
        total_unmatched_statement_debits:
          type: string
        total_unmatched_statement_credits:
          type: string
        unmatched_statement_count:
          type: integer
          format: int32
        outstanding_deposits:
          type: string
        outstanding_checks:
          type: string
        deposits_in_transit:
          type: string
        bank_charges:
          type: string
        bank_interest:
          type: string
        nsf_checks:
          type: string
        other_adjustments:
          type: string
        adjusted_book_balance:
          nullable: true
          type: string
        adjusted_statement_balance:
          nullable: true
          type: string
        difference:
          type: string
        is_balanced:
          type: boolean
        status:
          type: string
        started_at:
          nullable: true
          type: string
          format: date-time
        started_by:
          nullable: true
          type: string
          format: uuid
        completed_at:
          nullable: true
          type: string
          format: date-time
        completed_by:
          nullable: true
          type: string
          format: uuid
        reviewed_at:
          nullable: true
          type: string
          format: date-time
        reviewed_by:
          nullable: true
          type: string
          format: uuid
        has_adjusting_entries:
          type: boolean
        adjusting_journal_ids:
          type: object
        statement_source:
          nullable: true
          type: string
          maxLength: 50
        statement_file_url:
          nullable: true
          type: string
          maxLength: 500
        import_date:
          nullable: true
          type: string
          format: date-time
        notes:
          nullable: true
          type: string
          maxLength: 1000
        discrepancy_notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    ReconciliationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Reconciliation'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PartialReconcile:
      type: object
      required:
        - id
        - company_id
        - debit_move_id
        - credit_move_id
        - amount
        - currency
        - max_date
        - origin
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        debit_move_id:
          type: string
          format: uuid
        credit_move_id:
          type: string
          format: uuid
        full_reconcile_id:
          nullable: true
          type: string
          format: uuid
        exchange_move_id:
          nullable: true
          type: string
          format: uuid
        amount:
          type: string
        debit_amount_currency:
          nullable: true
          type: string
        credit_amount_currency:
          nullable: true
          type: string
        currency:
          type: string
          maxLength: 3
        max_date:
          type: string
          format: date
        origin:
          type: string
        source_type:
          nullable: true
          type: string
          maxLength: 50
        source_id:
          nullable: true
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          nullable: true
          type: object
...
    CreatePartialReconcileRequest:
      type: object
      required:
        - company_id
        - debit_move_id
        - credit_move_id
        - amount
        - currency
        - max_date
        - origin
      properties:
        company_id:
          type: string
          format: uuid
        debit_move_id:
          type: string
          format: uuid
        credit_move_id:
          type: string
          format: uuid
        full_reconcile_id:
          nullable: true
          type: string
          format: uuid
        exchange_move_id:
          nullable: true
          type: string
          format: uuid
        amount:
          type: string
        debit_amount_currency:
          nullable: true
          type: string
        credit_amount_currency:
          nullable: true
          type: string
        currency:
          type: string
          maxLength: 3
        max_date:
          type: string
          format: date
        origin:
          type: string
        source_type:
          nullable: true
          type: string
          maxLength: 50
        source_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          nullable: true
          type: object
...
    PartialReconcileList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PartialReconcile'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []